Wire in meta data for the previous display date
We want to include the display date prior to the last major change so
that users can be informed what the release date was prior to the
major change being made. Note that this code relies on the change note
only being set when absolutely necessary. The UI may need to be
updated such that entering a change note is only possible when the
change warrents it.